add Tutorials tests.
